Cleaning up MSVC warnings

+++ modules/kv_core/util/FileHelpers.cpp
@@ -59,7 +59,7 @@ int64 calculateFileHashCode (const File& file)
 
 bool overwriteFileWithNewDataIfDifferent (const File& file, const void* data, size_t numBytes)
 {
-    if (file.getSize() == numBytes
+    if (file.getSize() == static_cast<int> (numBytes)
           && calculateMemoryHashCode (data, numBytes) == calculateFileHashCode (file))
         return true;
